% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_dictionary.R
\name{bind_dictionary}
\alias{bind_dictionary}
\title{Bind Dictionaries Together}
\usage{
bind_dictionary(x, y, conflict_preference = NULL, keep_unmatched_y = FALSE)
}
\arguments{
\item{x, y}{A \link[=data_dictionary]{data dictionary} object.}

\item{conflict_preference}{a character value indicating what to do
when \code{x} and \code{y} have overlapping variables. If "left", then
the definitions in the \code{x} dictionary are preferred because
\code{x} is on the left of \code{y} in the function arguments. If "right",
\code{y} definitions are preferred.}
}
\value{
A \link[=data_dictionary]{data dictionary} object.
}
\description{
Binding dictionaries together can be useful when you have variables that
exist in multiple datasets and you don't want to re-write their meta
information into each of those datasets.
}
\examples{

dd_age_years <- data_dictionary(
  numeric_variable(
    name = "age_years",
    label = "Age",
    units = 'years',
    divby_modeling = 10
  )
)

dd_age_group <- data_dictionary(
  nominal_variable(
    name = "age_group",
    label = "Age group",
    description = "Ages of 0 to < 50, 50 to < 60, and \u2265 60 years",
    category_levels = c("age_lt_50", "age_gteq_50_lt_60", "age_gteq_60"),
    category_labels = c("0 to < 50", "50 to < 60", "\u2265 60")
  )
)

bind_dictionary(dd_age_years, dd_age_group)
}
