% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\name{data_dictionary}
\alias{data_dictionary}
\title{Create Data Dictionary from Variable Definitions}
\usage{
data_dictionary(..., .list = NULL, copy_on_modify = TRUE)
}
\arguments{
\item{...}{One or more objects inheriting from A \link[=numeric_variable]{numeric variable}
or A \link[=nominal_variable]{nominal variable}.}

\item{.list}{A list of data variables. This argument allows \code{data_dictionary} to be used programmatically and is optional. It is intended to be used as an alternative to \code{...}.}

\item{copy_on_modify}{a logical value indicating whether \code{set} functions
should modify the dictionary in place or copy it then modify. The
default is \code{TRUE} because dictionaries are almost always small
and trivial to copy.}
}
\value{
A \link[=data_dictionary]{data dictionary} object
}
\description{
Initializes a \code{DataDictionary} object from a set of variables.
Variables should be made using \code{\link[=numeric_variable]{numeric_variable()}} or \code{\link[=nominal_variable]{nominal_variable()}}
and passed directly as arguments.
}
\details{
Only one of  \code{...} and \code{.list} should be specified.
}
\examples{

age_years <- numeric_variable(
  name = "age",
  label = "Age of participant",
  units = "years"
)

gender <- nominal_variable(
  name = "gender",
  label = "Gender of participant",
  category_levels = c("M", "F"),
  category_labels = c("Male", "Female")
)

dd <- data_dictionary(age_years, gender)
print(dd)

dd <- data_dictionary(.list = list(age_years, gender))
print(dd)

}
