% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_unknowns}
\alias{get_unknowns}
\title{Get Unknown Fields in Dictionary}
\usage{
get_unknowns(
  dictionary = NULL,
  as_request = FALSE,
  as_code = FALSE,
  show_optional = FALSE
)
}
\arguments{
\item{dictionary}{A \link[=data_dictionary]{data dictionary} object}

\item{as_request}{logical. If \code{TRUE}, text is returned in a readable
format that can be pasted into other settings, e.g., an e-mail
requesting for the unknown information. If \code{FALSE} (the default),
unknowns are returned as a \code{tibble}.}

\item{as_code}{logical. If \code{TRUE}, text is returned as code that can
be copy/pasted into an R script. If \code{FALSE} (the default),
unknowns are returned as a \code{tibble}.}

\item{show_optional}{logical. If \code{TRUE}, unknowns will include all
possible fields of the variables in \code{dictionary}. If \code{FALSE} (the
default), only labels and units are presented.}
}
\value{
If \code{as_request = FALSE} and \code{as_code = FALSE},
a \link[tibble:tibble-package]{tibble} is returned.
If \code{as_request = TRUE} or \code{as_code = TRUE}, text is returned.
}
\description{
Get Unknown Fields in Dictionary
}
\examples{

# units isn't returned in the output because it is specified (known)
get_unknowns(data_dictionary(numeric_variable("a", units = 'years')))

# label isn't returned in the output because it is specified (known)
get_unknowns(data_dictionary(nominal_variable("b", label = 'example')))
dd <- as_data_dictionary(iris) \%>\%
  set_labels(Sepal.Length = "Sepal length",
            Sepal.Width = "Sepal width",
            Species = "Flower species") \%>\%
  set_descriptions(Species = "The species are all subtypes of iris") \%>\%
  set_units(Sepal.Length = "cm")

# as_request = TRUE returns this information in a shareable format
get_unknowns(dd, as_request = TRUE)

# as_code = TRUE gives a head start on filling in the missing info
get_unknowns(dd, as_code = TRUE)
}
