% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{identifier_variable}
\alias{identifier_variable}
\title{Create an Identifier Variable}
\usage{
identifier_variable(name, label = NULL, description = NULL)
}
\arguments{
\item{name}{Character. The name of the variable (required).}

\item{label}{Character. A short label for the variable (optional).}

\item{description}{Character. A longer description of the variable
(optional).}
}
\value{
An \code{IdentifierVariable} object.
}
\description{
Initializes an \code{IdentifierVariable} object, typically used for ID variables
like subject IDs.
}
\examples{
patient_id <- identifier_variable(
  name = "patient_id",
  label = "Patient Identifier"
)
patient_id
}
\seealso{
Other variable_constructors: 
\code{\link{date_variable}()},
\code{\link{nominal_variable}()},
\code{\link{numeric_variable}()}
}
\concept{variable_constructors}
