% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{numeric_variable}
\alias{numeric_variable}
\title{Create a Numeric Variable}
\usage{
numeric_variable(
  name,
  label = NULL,
  description = NULL,
  units = NULL,
  divby_modeling = NULL
)
}
\arguments{
\item{name}{Character. The name of the variable (required).}

\item{label}{Character. A short label for the variable (optional).}

\item{description}{Character. A longer description of the variable
(optional).}

\item{units}{Character. Units for the variable (e.g., "kg", "years")
(optional).}

\item{divby_modeling}{Numeric. A constant indicating what to divide the
variable by before modeling. This improves interpretability of model
coefficients (optional).}
}
\value{
A \code{NumericVariable} object.
}
\description{
Initializes a \code{NumericVariable} object, typically used to represent
continuous or quantitative data.
}
\examples{
age <- numeric_variable(
  name = "age",
  label = "Age of participant",
  units = "years",
  divby_modeling = 10
)
age
}
\seealso{
Other variable_constructors: 
\code{\link{date_variable}()},
\code{\link{identifier_variable}()},
\code{\link{logical_variable}()},
\code{\link{nominal_variable}()}
}
\concept{variable_constructors}
