% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_acronyms}
\alias{set_acronyms}
\title{Set and Remove Acronyms}
\usage{
set_acronyms(dictionary, ..., show_warnings = TRUE)
}
\arguments{
\item{dictionary}{A \link[=data_dictionary]{data dictionary} object}

\item{...}{one or more comma-separated name-value pairs. Pairs are of the
form of \code{acronym = "spelled-out acronym"} or \code{"acronym" = "spelled-out acronym"}.}

\item{show_warnings}{logical; whether to display warnings.}
}
\value{
A modified \code{dictionary}.
}
\description{
Add acronyms and their descriptions to a \code{DataDictionary}.
Acronyms can also be removed if they are no longer used.
}
\examples{
dd <- as_data_dictionary(iris)
dd$get_acronyms() # NULL

dd <- set_acronyms(dictionary = dd,
                   SBP = "systolic blood pressure",
                   DBP = "diastolic blood pressure")

dd$get_acronyms()
}
\seealso{
\code{\link{get_acronym_defs}}
}
