% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_default_dictionary}
\alias{set_default_dictionary}
\title{Set the default dictionary used by \code{perinary} functions}
\usage{
set_default_dictionary(dictionary)
}
\arguments{
\item{dictionary}{A \link[=data_dictionary]{data dictionary} object}
}
\description{
Set the default dictionary used by \code{perinary} functions
}
