% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_identifiers}
\alias{set_identifiers}
\title{Set Identifier Variables}
\usage{
set_identifiers(dictionary, ...)
}
\arguments{
\item{dictionary}{A \link[=data_dictionary]{data dictionary} object}

\item{...}{quoted or unquoted names of identifier variables}
}
\value{
a modified \code{dictionary}
}
\description{
Set Identifier Variables
}
