% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_label}
\alias{set_label}
\alias{set_description}
\alias{set_unit}
\alias{set_divby_modeling}
\alias{set_category_label}
\alias{set_category_order}
\title{Modify Data Dictionary Elements}
\usage{
set_label(dictionary, ...)

set_description(dictionary, ...)

set_unit(dictionary, ...)

set_divby_modeling(dictionary, ...)

set_category_label(dictionary, ..., modify_order = FALSE)

set_category_order(dictionary, ...)
}
\arguments{
\item{dictionary}{A \link[=data_dictionary]{data dictionary} object.}

\item{...}{Name-value pairs. The name gives the name of the variable
that will be modified. The value must be a:
\itemize{
\item character value for \code{set_label()}, \code{set_units()},
and \code{set_descriptions()}
\item numeric value for \code{set_divby_modeling()}
\item character vector for \code{set_category_levels()} and \code{set_category_labels()}
}

See examples for examples of each.}
}
\value{
a modified \code{dictionary}
}
\description{
Set metadata in a \code{DataDictionary} object.
}
\examples{

dd <- as_data_dictionary(data.frame(a = 1, b = "cat")) \%>\%
 set_label(a = "numeric example", b = "categorical example") \%>\%
 set_unit(a = "years") \%>\%
 set_divby_modeling(a = 10) \%>\%
 set_description(a = "A variable used for examples") \%>\%
 set_category_label(b = c("cat" = "A small lion"))

dd

}
