% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_labels}
\alias{set_labels}
\alias{set_descriptions}
\alias{set_units}
\alias{set_divby_modeling}
\alias{set_category_labels}
\alias{set_category_order}
\title{Modify Data Dictionary Elements}
\usage{
set_labels(dictionary, ...)

set_descriptions(dictionary, ...)

set_units(dictionary, ...)

set_divby_modeling(dictionary, ...)

set_category_labels(dictionary, ...)

set_category_order(dictionary, ...)
}
\arguments{
\item{dictionary}{A \link[=data_dictionary]{data dictionary} object.}

\item{...}{Name-value pairs. The name gives the name of the variable
that will be modified. The value must be a:
\itemize{
\item character value for \code{set_labels()}, \code{set_units()},
and \code{set_descriptions()}
\item numeric value for \code{set_divby_modeling()}
\item character vector for \code{set_category_levels()} and \code{set_category_labels()}
}

See examples for examples of each.}
}
\value{
a modified \code{dictionary}
}
\description{
Set metadata in a \code{DataDictionary} object.
}
\examples{

dd <- as_data_dictionary(data.frame(a = 1, b = "cat", id = 1)) \%>\%
 set_identifiers(id) \%>\%
 set_labels(a = "numeric example", b = "categorical example") \%>\%
 set_units(a = "years") \%>\%
 set_divby_modeling(a = 10) \%>\%
 set_descriptions(a = "A variable used for examples") \%>\%
 set_category_labels(b = c("cat" = "A small lion"))

dd

}
