% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_templates}
\alias{set_templates}
\alias{set_label_templates}
\alias{set_description_templates}
\title{Set Label and Description Templates}
\usage{
set_label_templates(dictionary, ..., show_duplicate_template_warnings = TRUE)

set_description_templates(
  dictionary,
  ...,
  show_duplicate_template_warnings = TRUE
)
}
\arguments{
\item{dictionary}{A \link[=data_dictionary]{data dictionary} object}

\item{...}{one or more formulas. See details.}

\item{show_duplicate_template_warnings}{logical; whether to display a warning
if multiple templates will be assigned to the same row of
\code{get_dictionary(dictionary)}.}
}
\value{
a modified \code{dictionary}.
}
\description{
Add label or description templates to a \code{DataDictionary}.
}
\examples{
library(perinary)

dd <- as_data_dictionary(iris)
dd

dd <- set_label_templates(
  dictionary = dd,
  # dimension will be either "length" or "width"
  contains("Sepal") ~ "Sepal {dimension}.",
  contains("Petal") ~ "Petal {dimension}."
)

dd # the label column has changed

dd <- set_description_templates(
  dictionary = dd,
  # dimension will be either "length" or "width"
  contains("Sepal") ~ "The {dimension} of the sepal, in centimeters.",
  contains("Petal") ~ "The {dimension} of the petal, in centimeters."
)

dd # the description column has changed

}
