% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_template}
\alias{use_template}
\title{Use a template to document a variable}
\usage{
use_template(...)
}
\arguments{
\item{...}{name-value pairs. The names must match with variables in the
dictionary that have a template.}
}
\value{
an environment
}
\description{
This function should only be used inside of \code{\link[=set_labels]{set_labels()}} or
\code{\link[=set_descriptions]{set_descriptions()}}. Calling \code{use_template()} in these functions will
allow you to plug in values for an existing template, which can be
set in a dictionary using \code{set_label_templates}
}
