% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{variable_constructors}
\alias{variable_constructors}
\title{Variable Constructors}
\arguments{
\item{name}{Character. The name of the variable (required).}

\item{label}{Character. A short label for the variable (optional).}

\item{description}{Character. A longer description of the variable
(optional).}
}
\description{
Constructors for variable metadata classes (\code{NumericVariable},
\code{NominalVariable}, \code{DateVariable}, \code{IdentifierVariable}).
}
\keyword{internal}
