% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDictionary.R
\name{as_data_dictionary}
\alias{as_data_dictionary}
\title{Create Data Dictionary from Existing Data}
\usage{
as_data_dictionary(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
A \link[=data_dictionary]{data dictionary} object.
}
\description{
Create Data Dictionary from Existing Data
}
\examples{

attr(iris$Species, 'label') <- "Flower species"

as_data_dictionary(iris)

}
