% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_acronym_defs}
\alias{get_acronym_defs}
\title{Get Acronym Definitions}
\usage{
get_acronym_defs(dictionary, acronyms = NULL, show_warnings = TRUE)
}
\arguments{
\item{dictionary}{A \link[=data_dictionary]{data dictionary} object}

\item{acronyms}{\code{NULL} or a character vector of acronyms. One or more of
\code{names(dictionary$get_acronyms())}. If \code{NULL}, all acronyms in \code{dictionary}
will be selected.}

\item{show_warnings}{logical; whether to display warnings.}
}
\value{
a character string of the form "acronym1 = acronym 1 description;
acronym2 = acronym 2 description."
}
\description{
Get nicely formatted acronym definitions. Primarily intended for
use in footnotes.
}
\examples{
dd <- as_data_dictionary(iris)
dd$get_acronyms() # NULL

dd <- set_acronyms(dictionary = dd,
                   BP = "blood pressure",
                   SBP = "systolic blood pressure",
                   DBP = "diastolic blood pressure")

dd$get_acronyms()

get_acronym_defs(dd, acronyms = NULL) # use all acronyms (default)
get_acronym_defs(dd, acronyms = c("DBP", "BP"))

# Attempt to select an acronym that was not set
get_acronym_defs(dd, acronyms = "FOO") # character(0L)

}
\author{
Tyler Sagendorf
}
