% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perinary-package.R
\docType{package}
\name{perinary-package}
\alias{perinary}
\alias{perinary-package}
\title{perinary: Data Dictionaries}
\description{
Keeps labels and other relevant information about variables in a lightweight class that can be applied to datasets in various ways, aiming to eliminate redundancies and promote readable, maintainable code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/perisphere-rwe/perinary}
  \item Report bugs at \url{https://github.com/perisphere-rwe/perinary/issues}
}

}
\author{
\strong{Maintainer}: Byron Jaeger \email{bcjaeger@perisphere-rwe.com} (\href{https://orcid.org/0000-0001-7399-2299}{ORCID})

Other contributors:
\itemize{
  \item Tyler Sagendorf [contributor]
}

}
\keyword{internal}
